refactor: extract location logic into LocationController
Follows the same architecture pattern as other controllers (search,
eventstate, bottomsheet, directions):

- Created ui/map/location/ subdirectory
- Moved LocationManager.kt to location/
- Created LocationController.kt with all location state and methods
- Updated MapScreenViewModel to delegate to LocationController
- Updated test imports

No functional changes - just architectural cleanup to match the
refactored main branch structure.
